+++ mock/control-api/Cargo.toml
@@ -20,7 +20,7 @@ clap = { version = "3.0", features = ["derive"] }
 derive_more = "0.99"
 dotenv = "0.15"
 humantime-serde = "1.0"
-medea-control-api-proto = { version = "0.8.0-dev", path = "../../proto/control-api", features = ["grpc"] }
+medea-control-api-proto = { version = "0.8.0-dev", path = "../../proto/control-api", features = ["grpc", "client"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 slog = "2.5"
