+++ proto/control-api/build.rs
@@ -131,19 +131,40 @@ mod grpc {
                 println!("cargo:rerun-if-changed={filename}");
             });
 
-        for filename in &out_files {
-            if let Err(err) = fs::File::open(filename) {
-                if err.kind() != io::ErrorKind::NotFound {
-                    return Err(err.to_string().into());
-                }
+        for (proto, out) in grpc_spec_files.iter().zip(&out_files) {
+            let build_client = (cfg!(feature = "client")
+                && out.ends_with("api.rs"))
+                || (cfg!(feature = "server") && out.ends_with("callback.rs"));
+            let build_server = (cfg!(feature = "client")
+                && out.ends_with("callback.rs"))
+                || (cfg!(feature = "server") && out.ends_with("api.rs"));
 
-                tonic_build::configure()
-                    .out_dir(GRPC_DIR)
-                    .build_client(cfg!(feature = "client"))
-                    .build_server(cfg!(feature = "server"))
-                    .compile(&grpc_spec_files, &[GRPC_DIR.to_owned()])?;
-            }
+            tonic_build::configure()
+                .out_dir(GRPC_DIR)
+                .build_client(build_client)
+                .build_server(build_server)
+                .compile(&[proto], &[GRPC_DIR.to_owned()])?;
         }
+
+        // Remove empty `google.protobuf.rs` file generated by
+        // `prost-build`. This file doesn't affect to any
+        // functionality. It's generated only because of
+        // `prost-build` bug.
+        //
+        // Read instrumentisto/medea#95, hyperium/tonic#314 and
+        // danburkert/prost#228 for more info.
+        fs::remove_file(format!("{}/google.protobuf.rs", GRPC_DIR)).map_err(
+            |e| {
+                format!(
+                    "`google.protobuf.rs` file isn't generated. \
+                     This is good news, because hyperium/tonic#314 \
+                     issue was probably really fixed. Check it and \
+                     if so, then just remove this line of code.\n\
+                     Error: {}",
+                    e,
+                )
+            },
+        )?;
 
         Ok(())
     }
