+++ proto/control-api/src/grpc/server.rs
@@ -3,6 +3,8 @@
 use std::collections::HashMap;
 
 use async_trait::async_trait;
+use derive_more::{Display, Error, From};
+use tonic::codegen::{Body, Bytes};
 
 use crate::{
     callback::Request as CallbackRequest,
@@ -12,14 +14,15 @@ use crate::{
             self as control_proto,
             control_api_server::ControlApi as GrpcControlApi,
         },
-        callback::{
-            self as callback_proto,
-            callback_server::Callback as GrpcCallbackApi,
-        },
-        ProtobufError,
+        callback as callback_proto, CallbackClient, ProtobufError,
     },
     CallbackApi, ControlApi, Ping,
 };
+
+/// [`Box`]ed [`Error`] with [`Send`] and [`Sync`].
+///
+/// [`Error`]: std::error::Error
+type StdError = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 #[async_trait]
 impl<T> GrpcControlApi for T
@@ -161,21 +164,36 @@ where
 }
 
 #[async_trait]
-impl<T> GrpcCallbackApi for T
+impl<T> CallbackApi for CallbackClient<T>
 where
-    T: CallbackApi + Send + Sync + 'static,
-    T::Error: From<ProtobufError>,
-    tonic::Status: From<T::Error>,
+    T: Clone + tonic::client::GrpcService<tonic::body::BoxBody> + Send + Sync,
+    T::Future: Send,
+    T::Error: Into<StdError>,
+    T::ResponseBody: Body<Data = Bytes> + Send + 'static,
+    <T::ResponseBody as Body>::Error: Into<StdError> + Send,
 {
-    async fn on_event(
-        &self,
-        req: tonic::Request<callback_proto::Request>,
-    ) -> Result<tonic::Response<callback_proto::Response>, tonic::Status> {
-        let req = CallbackRequest::try_from(req.into_inner())
-            .map_err(T::Error::from)?;
-        self.on_event(req)
+    type Error = CallbackClientError;
+
+    async fn on_event(&self, req: CallbackRequest) -> Result<(), Self::Error> {
+        // It's ok to `.clone()` here.
+        // https://docs.rs/tonic/latest/tonic/client/index.html#concurrent-usage
+        let mut this = self.clone();
+
+        Self::on_event(&mut this, callback_proto::Request::from(req))
             .await
-            .map(|_| tonic::Response::new(callback_proto::Response {}))
+            .map(drop)
             .map_err(Into::into)
     }
+}
+
+/// [`CallbackClient`] error.
+#[derive(Debug, Display, From, Error)]
+pub enum CallbackClientError {
+    /// gRPC server errored.
+    #[display(fmt = "gRPC server errored: {}", _0)]
+    Tonic(tonic::Status),
+
+    /// Failed to convert from protobuf.
+    #[display(fmt = "Failed to convert from protobuf: {}", _0)]
+    TryFromProtobufError(ProtobufError),
 }
